MF-1002 - Refactor things tracing files
