+++ things/tracing/things.go
@@ -13,7 +13,6 @@ import (
 )
 
 const (
-	saveThing                  = "save_thing"
 	saveThings                 = "save_things"
 	updateThing                = "update_thing"
 	updateThingGroupAndProfile = "update_thing_group_and_profile"
@@ -22,21 +21,13 @@ const (
 	retrieveThingsByProfile    = "retrieve_things_by_profile"
 	retrieveThingsByGroups     = "retrieve_things_by_groups"
 	removeThing                = "remove_thing"
-	removeKey                  = "remove_key"
-	retrieveThingIDByKey       = "retrieve_id_by_key"
 	retrieveAllThings          = "retrieve_all_things"
 	backupAllThings            = "backup_all_things"
-	saveGroupIDByThingID       = "save_group_id_by_thing_id"
-	retrieveGroupIDByThingID   = "retrieve_group_id_by_thing_id"
-	removeGroupIDByThingID     = "remove_group_id_by_thing_id"
 	updateExternalKey          = "update_external_key"
 	removeExternalKey          = "remove_external_key"
 )
 
-var (
-	_ things.ThingRepository = (*thingRepositoryMiddleware)(nil)
-	_ things.ThingCache      = (*thingCacheMiddleware)(nil)
-)
+var _ things.ThingRepository = (*thingRepositoryMiddleware)(nil)
 
 type thingRepositoryMiddleware struct {
 	tracer opentracing.Tracer
@@ -146,74 +137,4 @@ func (trm thingRepositoryMiddleware) RemoveExternalKey(ctx context.Context, thin
 	ctx = opentracing.ContextWithSpan(ctx, span)
 
 	return trm.repo.RemoveExternalKey(ctx, thingID)
-}
-
-type thingCacheMiddleware struct {
-	tracer opentracing.Tracer
-	cache  things.ThingCache
-}
-
-// ThingCacheMiddleware tracks request and their latency, and adds spans
-// to context.
-func ThingCacheMiddleware(tracer opentracing.Tracer, cache things.ThingCache) things.ThingCache {
-	return thingCacheMiddleware{
-		tracer: tracer,
-		cache:  cache,
-	}
-}
-
-func (tcm thingCacheMiddleware) Save(ctx context.Context, key things.ThingKey, thingID string) error {
-	span := dbutil.CreateSpan(ctx, tcm.tracer, saveThing)
-	defer span.Finish()
-	ctx = opentracing.ContextWithSpan(ctx, span)
-
-	return tcm.cache.Save(ctx, key, thingID)
-}
-
-func (tcm thingCacheMiddleware) ID(ctx context.Context, key things.ThingKey) (string, error) {
-	span := dbutil.CreateSpan(ctx, tcm.tracer, retrieveThingIDByKey)
-	defer span.Finish()
-	ctx = opentracing.ContextWithSpan(ctx, span)
-
-	return tcm.cache.ID(ctx, key)
-}
-
-func (tcm thingCacheMiddleware) RemoveThing(ctx context.Context, thingID string) error {
-	span := dbutil.CreateSpan(ctx, tcm.tracer, removeThing)
-	defer span.Finish()
-	ctx = opentracing.ContextWithSpan(ctx, span)
-
-	return tcm.cache.RemoveThing(ctx, thingID)
-}
-
-func (tcm thingCacheMiddleware) RemoveKey(ctx context.Context, key things.ThingKey) error {
-	span := dbutil.CreateSpan(ctx, tcm.tracer, removeKey)
-	defer span.Finish()
-	ctx = opentracing.ContextWithSpan(ctx, span)
-
-	return tcm.cache.RemoveKey(ctx, key)
-}
-
-func (tcm thingCacheMiddleware) SaveGroup(ctx context.Context, thingID string, groupID string) error {
-	span := dbutil.CreateSpan(ctx, tcm.tracer, saveGroupIDByThingID)
-	defer span.Finish()
-	ctx = opentracing.ContextWithSpan(ctx, span)
-
-	return tcm.cache.SaveGroup(ctx, thingID, groupID)
-}
-
-func (tcm thingCacheMiddleware) ViewGroup(ctx context.Context, thingID string) (string, error) {
-	span := dbutil.CreateSpan(ctx, tcm.tracer, retrieveGroupIDByThingID)
-	defer span.Finish()
-	ctx = opentracing.ContextWithSpan(ctx, span)
-
-	return tcm.cache.ViewGroup(ctx, thingID)
-}
-
-func (tcm thingCacheMiddleware) RemoveGroup(ctx context.Context, thingID string) error {
-	span := dbutil.CreateSpan(ctx, tcm.tracer, removeGroupIDByThingID)
-	defer span.Finish()
-	ctx = opentracing.ContextWithSpan(ctx, span)
-
-	return tcm.cache.RemoveGroup(ctx, thingID)
 }
